small MP changes

+++ src/Multiplayer/Request.java
@@ -1,8 +1,8 @@
 package Multiplayer;
 
 public class Request {
-	public static final int CLOSE_CONNECTION = 1;
-	public static final int CHUNK_DATA = 2;
-	public static final int PLAYER_DATA = 3;
-	public static final int MAP_DATA = 4;
+	public static final byte CLOSE_CONNECTION = 1;
+	public static final byte CHUNK_DATA = 2;
+	public static final byte PLAYER_DATA = 3;
+	public static final byte MAP_DATA = 4;
 }
